[front] fix: move the folder `mocks` out of src/service...
...to fix the build on staging and production.
