+++ frontend/src/mocks/TalkEntry.ts
